Refactor: rename webapp to singleplayer
