+++ src/singleplayer/api.py
