+++ src/test/test_websockets.py
@@ -119,7 +119,7 @@ def test_join_game_different_difficulty(test_clients):
     assert not r2[1]["args"][0]['ready']
 
 
-@patch('webapp.api.classifier', mock_classifier)
+@patch('singleplayer.api.classifier', mock_classifier)
 def test_classification_only_client1_correct(test_clients):
     time_left = 1
     correct_label = "angel"
@@ -160,7 +160,7 @@ def test_classification_only_client1_correct(test_clients):
     assert len(r2) == 1
 
 
-@patch('webapp.api.classifier', mock_classifier)
+@patch('singleplayer.api.classifier', mock_classifier)
 def test_game_in_different_languages(test_clients):
     correct_label = "angel"
     wrong_label = "bicycle"
@@ -203,7 +203,7 @@ def test_game_in_different_languages(test_clients):
     assert len(r2) == 1
 
 
-@patch('webapp.api.classifier', mock_classifier)
+@patch('singleplayer.api.classifier', mock_classifier)
 def test_classification_both_correct(test_clients):
     time_left = 1
     correct_label = "angel"
@@ -252,7 +252,7 @@ def test_classification_both_correct(test_clients):
     assert len(r2) == 2
 
 
-@patch('webapp.api.classifier', mock_classifier)
+@patch('singleplayer.api.classifier', mock_classifier)
 def test_classification_client1_timeout_and_client2_correct(test_clients):
     time_out = 0
     time_left = 1
@@ -297,7 +297,7 @@ def test_classification_client1_timeout_and_client2_correct(test_clients):
     assert len(r2) == 2
 
 
-@patch('webapp.api.classifier', mock_classifier)
+@patch('singleplayer.api.classifier', mock_classifier)
 def test_classification_client1_correct_and_client2_timeout(test_clients):
     time_out = 0
     time_left = 1
@@ -343,7 +343,7 @@ def test_classification_client1_correct_and_client2_timeout(test_clients):
     assert len(r2) == 1
 
 
-@patch('webapp.api.classifier', mock_classifier)
+@patch('singleplayer.api.classifier', mock_classifier)
 def test_classification_both_timeout(test_clients):
     time_out = 0
     correct_label = "angel"
